% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tables.R
\name{capwords}
\alias{capwords}
\title{Capitalize the first letter of each word in a character string.}
\usage{
capwords(s)
}
\arguments{
\item{s}{A character vector, or an object that can be coerced to character by
\code{\link[base]{as.character}}.}
}
\value{
A string where the first letter of each word is capitalized, and the
  remaining letters are lower-case.
}
\description{
\code{capwords} is a slightly simplified version of the \code{capwords}
  example function provided in the \code{\link[base]{toupper}} documentation.
}
\examples{
capwords("the quick brown fox did a thing.")
capwords("THE QUICK BROWN FOX DID A THING.")
}
