% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input}
\alias{check_input}
\title{Check whether a user input is a time or a comment.}
\usage{
check_input(input)
}
\arguments{
\item{input}{A numeric or charater string.}
}
\value{
Returns \code{TRUE} if the input could be coerced to a time,
  \code{FALSE} otherwise.
}
\description{
\code{check_input} takes a user's input and determines whether or not it can
  be coerced to a time via \code{format_time}.
}
