% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_time.R
\name{dm_date}
\alias{dm_date}
\title{Change System Date to Google Sheets date.}
\usage{
dm_date(date = Sys.Date(), shift = 0)
}
\arguments{
\item{date}{an optional parameter to specify the date in a YYY-MM-DD format.
Default is \code{Sys.Date()}.}

\item{shift}{An integer to define the number of days by which the data entry
date should be shifted. Default is 0. The day after today (tomorrow) is +1,
yesterday is -1, etc.}
}
\value{
The function returns a character string of the re-formatted date. For
  example, the \code{Sys.Date()} output on April 2, 2017 is
  \code{"2017-04-02"}, and the newly returned character string would be
  \code{"Sun, Apr 02"}.
}
\description{
\code{dm_date} converts the output from R's \code{Sys.date()} to the format
  currently used in our Daily Mini Google Spreadsheet. This spreadsheet
  format uses an abbreviated day of the week, an abbreviated month, and the
  numeric day of the month.
}
\examples{
dm_date()    # today
dm_date(shift = 1)   # tomorrow
dm_date(shift = -7)  # one week ago
}
