% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_time.R
\name{edit_data}
\alias{edit_data}
\title{Write a value to a Google Sheet.}
\usage{
edit_data(ss, d, input, user, unames, date = Sys.Date(), shift = 0)
}
\arguments{
\item{ss}{a registered Google spreadsheet, i.e. a \code{\link{googlesheet}}
object}

\item{d}{A data frame version of the "Data" worksheet in spreadsheet
\code{ss} (see \code{\link[googlesheets]{googlesheet}}).}

\item{input}{new cell values, as an object that can be coerced into a
character vector, presumably an atomic vector, a factor, a matrix or a
data.frame}

\item{user}{A character string providing a username, which should
correspond to a column in the spreadsheet.}

\item{unames}{A user-defined \code{\link[base]{switch}} function that
identifies user-specific column-names given a username (see
\code{\link{write_time}} examples).}

\item{date}{an optional parameter to specify the date in a YYY-MM-DD format.
Default is \code{Sys.Date()}.}

\item{shift}{An integer to define the number of days by which the data entry
date should be shifted. Default is 0. The day after today (tomorrow) is +1,
yesterday is -1, etc.}
}
\description{
\code{edit_data} is a wrapper for \code{\link[googlesheets]{gs_edit_cells}}
  that writes \code{input} to a worksheet called 'Data', in a cell specified
  by \code{user} and \code{shift}.
}
\details{
This function is not called directly, but is a helper function for
  \code{\link{write_time}}.
}
