% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_time.R
\name{format_time}
\alias{format_time}
\title{Standardize some input time to an HH:MM:SS format.}
\usage{
format_time(time)
}
\arguments{
\item{time}{Either a single numeric, corresponding to some number of seconds,
or a character string where hours, minutes, and/or seconds are separated by
a colon (\code{:}).}
}
\value{
a character string of the form \code{HH:MM:SS}.
}
\description{
\code{format_time} takes either a numeric or character input \code{s},
  corresponding to some time, and returns a string that is formatted to a
  consistent \code{HH:MM:SS} format.
}
\details{
\code{format_time} is flexible with regard to the format of \code{time},
  but if hours and/or minutes are supplied in addition to seconds, they must
  be some form where mintues precede seconds, and hours precede minutes.
  Thus, to indicate 1 hour, 30 minutes, 5 seconds, the following values of
  \code{time} are all acceptable: \code{"1:30:05"}, \code{5405},
  \code{"5405"}, \code{"30:3605"}, \code{"1.5:0:5"}. However,
  \code{1.5:05}, is not.
}
\examples{
format_time(2.5) # fractions are okay

# Many different ways to express "01:30:05":
format_time("1:30:05")  # 1 hour, 30 minutes, 5 seconds
format_time(5405)       # 5405 seconds (numeric)
format_time("5405")     # 5405 seconds (string)
format_time("30:3605")  # 30 minutes, 3605 seconds
format_time("1.5:0:5")  # 1.5 hours, 0 minutes, 5 seconds

# Order is important
format_time("1.5:5")    # 1.5 minutes, 5 seconds
}
