% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_time.R
\name{get_cell}
\alias{get_cell}
\title{Determine cell location in spreadsheet.}
\usage{
get_cell(d, user_col, date = Sys.Date(), shift = 0)
}
\arguments{
\item{d}{A data frame version of the "Data" worksheet in spreadsheet
\code{ss} (see \code{\link[googlesheets]{googlesheet}}).}

\item{user_col}{A character string providing a username, which should
correspond to a column in the spreadsheet.}

\item{date}{an optional parameter to specify the date in a YYY-MM-DD format.
Default is \code{Sys.Date()}.}

\item{shift}{An integer to define the number of days by which the data entry
date should be shifted. Default is 0. The day after today (tomorrow) is +1,
yesterday is -1, etc.}
}
\value{
The function returns the alpha-numeric location of the cell (e.g,
  \code{"D42"}) as a character string.
}
\description{
\code{get_cell} finds a cell specified by a column that corresponds to
  \code{user_col} and a row specified by today's date, possibly
  shifted by \code{shift}.
}
\details{
This function is not called directly, but is a helper function for
  \code{\link{write_time}}.
}
