% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ws}
\alias{get_ws}
\title{Silently connect to and download a Google Sheet.}
\usage{
get_ws(key, ws)
}
\arguments{
\item{key}{Path to a file that contains s unique key for the Google sheet.
Should be readable by \code{link[base]{scan}} with \code{what="char"}.}

\item{ws}{A character string specifying title of the worksheet.}
}
\value{
The function returns a list. The first element of the list, \code{ss}
  is a googlesheet object, the second element, \code{ww}, is a tibble of the
  worksheet.
}
\description{
\code{get_ws} takes a \code{key} and a worksheet (\code{ws}) and silently
  returns the googlesheet object
  (see \code{\link[googlesheets]{googlesheet}}) and a tibble of the
  worksheet.
}
