% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hms_to_s}
\alias{hms_to_s}
\title{Convert an HH:MM:SS formatted string to seconds.}
\usage{
hms_to_s(x)
}
\arguments{
\item{x}{A character string of the form \code{HH:MM:SS}.}
}
\value{
The number of seconds (numeric) represented by \code{HH:MM:SS}.
}
\description{
\code{hms_to_s} takes a character string formatted in \code{HH:MM:SS} and
  returns the equivalent number of seconds.
}
\examples{
hms_to_s(10:10:10)
hms_to_s(00:10:10)
hms_to_s(00:00:10)
}
