% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tables.R
\name{print_board}
\alias{print_board}
\title{Print nicely formatted Rankings Tables.}
\usage{
print_board(tab, r, given_only = TRUE)
}
\arguments{
\item{tab}{A character input for the table to be displayed. Can be any of the
following:
 \itemize{
   \item Average Times: \code{"AVERAGES"}, \code{"AVG"}, or \code{"A"}
   \item Personal Bests: \code{"BESTS"}, \code{"BEST"}, or \code{"B"}
   \item Games Played: \code{"GAMES"}, \code{"GAME"}, or \code{"G"}
   \item Monthly Leaderboard: \code{"MONTH"}, \code{"MO"}, or \code{"M"}
   \item All-Time Points: \code{"POINTS"}, \code{"POINT"}, or \code{"P"}
   \item All-Time Wins: \code{"WINS"}, \code{"WIN"}, or \code{"W"}
 }}

\item{r}{A data frame version of the "Ranks" worksheet in the Daily Mini
spreadsheet.}

\item{given_only}{A boolean that determines whether or not to only print the
player's given name in the table. Defaults to \code{TRUE}.}
}
\value{
A formatted table.
}
\description{
\code{print_board} takes a table given by \code{tab} and prints a
  simply-formatted version of that table. Tables are narrow enough to be
  displayed in the Slack app on an iPhone 6.
}
