% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tables.R
\name{print_scoreboard}
\alias{print_scoreboard}
\title{Print nicely formatted Scoreboard}
\usage{
print_scoreboard(s, given_only = TRUE)
}
\arguments{
\item{s}{A data frame version of the "simple_dashboard" worksheet in the
Daily Mini spreadsheet.}

\item{given_only}{A boolean that determines whether or not to only print the
player's given name in the table. Defaults to \code{TRUE}.}
}
\value{
A formatted table.
}
\description{
\code{print_scoreboard} prints a simply-formatted version of the 'Dashboard'
  worksheet from the Daily Mini spreadsheet. Tables are narrow enough to be
  displayed in the Slack app on an iPhone 6.
}
\details{
Technically, this function prints a simply-formatted version of the
  'simple_dashboard' worksheet. This worksheet is nearly identical to the
  'Dashboard' worksheet apart from some labeling and formatting differences.
  One major difference, however, is that 'Dashboard' sometimes reports
  multiple statistics for a single category (e.g., 'Total Points' and
  'Average Time'), while 'simple_dashboard' only reports the most relevant.
}
