% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tables.R
\name{rank_list}
\alias{rank_list}
\title{Generate a list of the Rankings Tables.}
\usage{
rank_list(r)
}
\arguments{
\item{r}{A data frame version of the "Ranks" worksheet in the Daily Mini
spreadsheet.}
}
\value{
A list of length 6, where each element of the list returns the
  corresponding rankings table.
}
\description{
Given a dataframe of all of the "Ranks" worksheet, \code{rank_list} generates
  a list where each element of the list is a nicely-formatted table. Each
  table can then be called using the syntax \code{r_list[[table]]}.
}
\details{
This function is not called directly, but is instead a helper function for
  \code{link{print_board}}
}
