% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_int}
\alias{round_int}
\title{Round down to the nearest integer.}
\usage{
round_int(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The function returns the vector \code{x} rounded to the nearest
  integer.
}
\description{
\code{round_int} takes a numeric input \code{x} and rounds each element
  to the nearest integer. Ties are broken using a 'half-up' rule. This is
  different from R's default rounding method, which breaks ties by rounding
  to the nearest even number (i.e. 'half-even').
}
\examples{
round_int(2.5)  # rounding 'half-up' returns 3
round(2.5)      # rounding 'half-even', R's default, return 2
}
