% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s_to_hms}
\alias{s_to_hms}
\title{Convert some number of seconds into an HH:MM:SS format.}
\usage{
s_to_hms(x)
}
\arguments{
\item{x}{A single number, in seconds.}
}
\value{
a character string of the form \code{HH:MM:SS}.
}
\description{
\code{s_to_hms} takes a numeric input \code{s} (or an input that can be
  coerced to numeric) that corresponds to some number of seconds, and returns
  a string that is formatted to the appropriate \code{HH:MM:SS}.
}
\examples{
s_to_hms(2.5)
s_to_hms(120)
s_to_hms(4242)
}
