% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_message}
\alias{slack_message}
\title{Given inputs, uses httr::POST to post a message to Slack.}
\usage{
slack_message(url, channel = NULL, username = NULL, text = NULL,
  private = TRUE)
}
\arguments{
\item{url}{The \code{return_url} provided by Slack (character string).}

\item{channel}{The channel to post the message to (character string).}

\item{username}{The username to post in response to (character string).}

\item{text}{The text that should be posted (character string).}

\item{private}{Whether this message should be posted privately to the user
(\code{TRUE}), or posted publically (\code{FALSE}).}
}
\description{
Given inputs, uses httr::POST to post a message to Slack.
}
