% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_parse}
\alias{slack_parse}
\title{A function to parse Slack input.}
\usage{
slack_parse(text, user_name)
}
\arguments{
\item{text}{Input from Slack.}

\item{user_name}{The username provided by Slack.}
}
\value{
A list containing variables required for \code{link{write_time}}.
}
\description{
Given text and a user_name, returns a list of variables needed for
  \code{link{write_time}}.
}
