% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_post_time}
\alias{slack_post_time}
\title{A function to return time messages.}
\usage{
slack_post_time(user_given, time, shift)
}
\arguments{
\item{user_given}{A user's first (given) name.}

\item{time}{Either an integer corresponding to a number of seconds elapsed
(e.g., 42 or 142), or a character string in the form:
\itemize{
  \item HH:MM:SS (e.g. \code{"01:23:45"})
  \item MM:SS    (e.g. \code{"23:45"})
  \item SS       (e.g. \code{"45"})
}
Any any number of hours, minutes, or seconds are valid (i.e.,
\code{"0:03"}, \code{"000:3"}, \code{"3"}, and \code{"00:00:03"} are all
valid ways to input three seconds).}

\item{shift}{An integer to define the number of days by which the data entry
date should be shifted. Default is 0. The day after today (tomorrow) is +1,
yesterday is -1, etc.}
}
\value{
A bespoke message
}
\description{
These messages are displayed publically so everyone knows what time the user
  input.
}
