% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_text_fail}
\alias{slack_text_fail}
\title{A function to return bespoke FAIL messages.}
\usage{
slack_text_fail(user_given)
}
\arguments{
\item{user_given}{A user's first (given) name.}
}
\value{
A random bespoke message that (often) includes the user's name.
}
\description{
When somebody fails, it hurts just a little bit. These messages are for them.
}
