% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_text_fails}
\alias{slack_text_fails}
\title{When somebody fails, it hurts just a little bit. These messages are for them.}
\usage{
slack_text_fails(u)
}
\arguments{
\item{u}{A user's name.}
}
\value{
A bespoke message
}
\description{
When somebody fails, it hurts just a little bit. These messages are for them.
}
