% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_text_pass}
\alias{slack_text_pass}
\title{A function to return bespoke SUCCESS message.}
\usage{
slack_text_pass(user_given, time)
}
\arguments{
\item{user_given}{A user's first (given) name.}

\item{time}{Either an integer corresponding to a number of seconds elapsed
(e.g., 42 or 142), or a character string in the form:
\itemize{
  \item HH:MM:SS (e.g. \code{"01:23:45"})
  \item MM:SS    (e.g. \code{"23:45"})
  \item SS       (e.g. \code{"45"})
}
Any any number of hours, minutes, or seconds are valid (i.e.,
\code{"0:03"}, \code{"000:3"}, \code{"3"}, and \code{"00:00:03"} are all
valid ways to input three seconds).}
}
\value{
A random bespoke message that (often) includes the user's name.
}
\description{
These people completed the puzzle for the day. Let's make them feel good!
}
