% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssplit}
\alias{ssplit}
\title{Splits a string.}
\usage{
ssplit(x, split)
}
\arguments{
\item{x}{A character string.}

\item{split}{character containing a regular expression that can be used for
splitting.}
}
\value{
Returns a character vector.
}
\description{
\code{ustrsplit} is a wrapper for \code{[base]{strsplit}} that only accepts a
  single character string, but in exchange, retuns a vector instead of a
  list.
}
