% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_time.R
\name{write_time}
\alias{write_time}
\title{Write a Daily Mini time to the Google Sheet.}
\usage{
write_time(ss, d, time, user, unames, date = Sys.Date(), shift = 0,
  overwrite = FALSE)
}
\arguments{
\item{ss}{a googlesheet object (see \code{\link[googlesheets]{googlesheet}}).}

\item{d}{A data frame version of the "Data" worksheet in spreadsheet
\code{ss} (see \code{\link[googlesheets]{googlesheet}}).}

\item{time}{Either an integer corresponding to a number of seconds elapsed
(e.g., 42 or 142), or a character string in the form:
\itemize{
  \item HH:MM:SS (e.g. \code{"01:23:45"})
  \item MM:SS    (e.g. \code{"23:45"})
  \item SS       (e.g. \code{"45"})
}
Any any number of hours, minutes, or seconds are valid (i.e.,
\code{"0:03"}, \code{"000:3"}, \code{"3"}, and \code{"00:00:03"} are all
valid ways to input three seconds).}

\item{user}{A character string providing a username, which should
correspond to a column in the spreadsheet.}

\item{unames}{A user-defined \code{\link[base]{switch}} function that
identifies user-specific column-names given a username (see
\code{\link{write_time}} examples).}

\item{date}{an optional parameter to specify the date in a YYY-MM-DD format.
Default is \code{Sys.Date()}.}

\item{shift}{An integer to define the number of days by which the data entry
date should be shifted. Default is 0. The day after today (tomorrow) is +1,
yesterday is -1, etc.}

\item{overwrite}{A boolean to determine whether existing values in the cell
should be overwritten. Default is \code{FALSE}.}
}
\value{
The function returns a message, stating which cell was written to,
  the value that was written, and if the cell was overwritten, a message
  to inform the user that the previous value was overwritten.
}
\description{
\code{write_time} takes a time provided by a specified user and writes that
  time to the corresponding cell in a data spreadsheet given by \code{ss}.
}
\examples{
\dontrun{

unames <- function(user) {
  switch(user,
         user_1 = "A_User",
         user_2 = "B_User",
         user_3 = "C_User"
  )
}

write_time(ss, d, "0:42", "user_1", unames = unames)
}
}
